Moved dns configuration attributes to own attribute and extracted nopoll and noresolv attributes out.
To be able to "not set" no-poll and no-resolv attributes there wasn't way
to do it previously. After this commit you can set nopoll and noresolv
attributes to false and it will not add those lines to conf -file.

Configuration have moved from node[:dnsmasq][:dns]
to node[:dnsmasq][:dns][:config] so this is not backward compatible.